Move setUserAccess and transferOwnership to ApiDomainController
